refactor: Move classes into a new subpackage for expressions
