+++ ProfCalculator/src/de/uulm/sp/swt/profcalculator/ProfCalculator.java
@@ -1,5 +1,9 @@
 package de.uulm.sp.swt.profcalculator;
 
+import de.uulm.sp.swt.profcalculator.expressions.Addition;
+import de.uulm.sp.swt.profcalculator.expressions.Expression;
+import de.uulm.sp.swt.profcalculator.expressions.Multiplication;
+import de.uulm.sp.swt.profcalculator.expressions.Value;
 import javafx.application.Application;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
