+++ ProfCalculator/src/de/uulm/sp/swt/profcalculator/expressions/Multiplication.java
@@ -1,4 +1,4 @@
-package de.uulm.sp.swt.profcalculator;
+package de.uulm.sp.swt.profcalculator.expressions;
 
 public class Multiplication extends Expression {
